requirements.txt
ultralytics>=8.0.0
streamlit>=1.30.0
Pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0
torch>=2.0.0
torchvision>=0.15.0
matplotlib>=3.7.0
pyyaml>=6.0
requests>=2.31.0
tqdm>=4.65.0
pydantic>=2.0.0
